Posts CRUD

+++ backend/src/main/java/com/basicsocialmedia/backend/controller/UserController.java
@@ -2,8 +2,12 @@ package com.basicsocialmedia.backend.controller;
 
 import com.basicsocialmedia.backend.model.User;
 import com.basicsocialmedia.backend.service.UserService;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,22 +20,22 @@ public class UserController {
 
     public UserController(UserService userService) { this.userService = userService; }
 
-    // GET /api/users(?username="")(?email="")
+    // GET /api/users(?username="")(?email="")&page=1&size=20
     @GetMapping
     public ResponseEntity<?> getUsersByQueryParam(
             @RequestParam(required = false) String username,
-            @RequestParam(required = false) String email) {
-        if (username != null) {
+            @RequestParam(required = false) String email, Pageable pageable) {
+        if (username != null && !username.isBlank() && (email == null || email.isBlank())) {
             Optional<User> user = userService.getUserByUsername(username);
             return user.map(u -> new ResponseEntity<>(u, HttpStatus.OK))
                     .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
         }
-        if (email != null) {
+        if (email != null && !email.isBlank() && (username == null || username.isBlank())) {
             Optional<User> user = userService.getUserByEmail(email);
             return user.map(u -> new ResponseEntity<>(u, HttpStatus.OK))
                     .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
         }
-        List<User> allUsers = userService.getAllUsers();
+        Page<User> allUsers = userService.getAllUsers(pageable);
         return new ResponseEntity<>(allUsers, HttpStatus.OK);
     }
 
@@ -51,15 +55,40 @@ public class UserController {
     }
 
     // PATCH /api/users/{id}
-    @PatchMapping("/{id}")
-    public ResponseEntity<User> patchUser(@PathVariable Integer id, User user) {
+    @PatchMapping("/{id}") // user can patch their own info
+    public ResponseEntity<User> patchUser(@PathVariable Integer id,@RequestBody User user) {
+        User requestedUser = userService.getUserById(id)
+                .orElseThrow(() -> new RuntimeException("User not found with the id: "+id));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUsername = authentication.getName();
+
+        User currentUser = userService.getUserByUsername(currentUsername)
+                .orElseThrow(() -> new RuntimeException("Authenticated user not found"));
+
+        if (!currentUser.getId().equals(requestedUser.getId())) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         User patchedUser = userService.patchUser(id, user);
         return new ResponseEntity<>(patchedUser, HttpStatus.OK);
     }
 
     // DELETE /api/users/{id}
-    @DeleteMapping("/{id}")
+    @DeleteMapping("/{id}") // user can delete their own account
     public ResponseEntity<Void> deleteUser(@PathVariable Integer id) {
+        User requestedUser = userService.getUserById(id)
+                        .orElseThrow(() -> new RuntimeException("User not found with id: "+id));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUsername = authentication.getName();
+        User currentUser = userService.getUserByUsername(currentUsername)
+                .orElseThrow(() -> new RuntimeException("Authenticated user not found"));
+
+        if (!currentUser.getId().equals(requestedUser.getId())) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         userService.deleteUser(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
